Revert "Add methods for processing JSON features in Gff3HandlerService for GFF3 export (#1615)"
This reverts commit f608ca29e93885f0690fa48a1801a2fffec6bdff.

+++ src/groovy/org/bbop/apollo/GFF3Entry.groovy
@@ -7,14 +7,14 @@ public class GFF3Entry {
     private String seqId;
     private String source;
     private String type;
-    private long start;
-    private long end;
+    private int start;
+    private int end;
     private String score;
     private String strand;
     private String phase;
     private Map<String, String> attributes;
 
-    public GFF3Entry(String seqId, String source, String type, long start, long end, String score, String strand, String phase) {
+    public GFF3Entry(String seqId, String source, String type, int start, int end, String score, String strand, String phase) {
         this.seqId = seqId;
         this.source = source;
         this.type = type;
@@ -50,19 +50,19 @@ public class GFF3Entry {
         this.type = type;
     }
 
-    public long getStart() {
+    public int getStart() {
         return start;
     }
 
-    public void setStart(long start) {
+    public void setStart(int start) {
         this.start = start;
     }
 
-    public long getEnd() {
+    public int getEnd() {
         return end;
     }
 
-    public void setEnd(long end) {
+    public void setEnd(int end) {
         this.end = end;
     }
